Move module `new_iter` to `iter`
